requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-appconf==1.0.4
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-filebrowser==3.14.1
django-grappelli==2.15.1
django-mathfilters==1.0.0
django-redis==5.0.0
django-select2==7.7.1
django-tempus-dominus==5.1.2.16
django-tinymce==3.3.0
djangorestframework==3.12.4
fpdf==1.7.2
gunicorn==20.1.0
idna==3.2
imgkit==1.2.2
jsonfield==3.1.0
pdfkit==0.6.1
Pillow==10.1.0
psycopg2==2.9.9
pyodbc==5.0.1
python-barcode==0.13.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
